+++ packages/headless/src/features/facets/facet-order/facet-order-slice.ts
@@ -6,11 +6,12 @@ import {getFacetOrderInitialState} from './facet-order-state';
 export const facetOrderReducer = createReducer(
   getFacetOrderInitialState(),
   (builder) => {
-    builder.addCase(executeSearch.fulfilled, (_, action) => {
-      return action.payload.response.facets.map((facet) => facet.facetId);
-    });
-    builder.addCase(change.fulfilled, (_, action) => {
-      return action.payload.facetOrder;
-    });
+    builder
+      .addCase(executeSearch.fulfilled, (_, action) => {
+        return action.payload.response.facets.map((facet) => facet.facetId);
+      })
+      .addCase(change.fulfilled, (state, action) => {
+        return action.payload?.facetOrder ?? state;
+      });
   }
 );