+++ packages/headless/src/features/history/history-state.ts
@@ -4,6 +4,7 @@ import {getContextInitialState} from '../context/context-state';
 import {getDebugInitialState} from '../debug/debug-state';
 import {getFacetOptionsInitialState} from '../facet-options/facet-options-state';
 import {getCategoryFacetSetInitialState} from '../facets/category-facet-set/category-facet-set-state';
+import {getFacetOrderInitialState} from '../facets/facet-order/facet-order-state';
 import {getFacetSetInitialState} from '../facets/facet-set/facet-set-state';
 import {getDateFacetSetInitialState} from '../facets/range-facets/date-facet-set/date-facet-set-state';
 import {getNumericFacetSetInitialState} from '../facets/range-facets/numeric-facet-set/numeric-facet-set-state';
@@ -18,28 +19,29 @@ export interface HistoryState extends SearchParametersState {
   facetOrder: string[];
 }
 
-function getSearchParametersInitialState(): SearchParametersState {
+export function getHistoryInitialState(): HistoryState {
   return {
     context: getContextInitialState(),
     facetSet: getFacetSetInitialState(),
-    dateFacetSet: getDateFacetSetInitialState(),
     numericFacetSet: getNumericFacetSetInitialState(),
+    dateFacetSet: getDateFacetSetInitialState(),
     categoryFacetSet: getCategoryFacetSetInitialState(),
-    facetOptions: getFacetOptionsInitialState(),
     pagination: getPaginationInitialState(),
     query: getQueryInitialState(),
     advancedSearchQueries: getAdvancedSearchQueriesInitialState(),
-    sortCriteria: getSortCriteriaInitialState(),
     querySet: getQuerySetInitialState(),
+    sortCriteria: getSortCriteriaInitialState(),
     pipeline: getPipelineInitialState(),
     searchHub: getSearchHubInitialState(),
+    facetOptions: getFacetOptionsInitialState(),
+    facetOrder: getFacetOrderInitialState(),
     debug: getDebugInitialState(),
   };
 }
 
-export function getHistoryInitialState(): HistoryState {
+export function extractHistory(state: Partial<HistoryState>): HistoryState {
   return {
-    ...getSearchParametersInitialState(),
-    facetOrder: [],
+    ...getHistoryInitialState(),
+    ...state,
   };
 }