+++ packages/headless/src/features/query/query-slice.ts
@@ -15,7 +15,10 @@ export const queryReducer = createReducer(getQueryInitialState(), (builder) =>
     .addCase(selectQuerySuggestion, (state, action) => {
       state.q = action.payload.expression;
     })
-    .addCase(change.fulfilled, (_, action) => action.payload.query)
+    .addCase(
+      change.fulfilled,
+      (state, action) => action.payload?.query ?? state
+    )
     .addCase(restoreSearchParameters, (state, action) => {
       state.q = action.payload.q ?? state.q;
       state.enableQuerySyntax =